+++ src/main/java/com/moplus/moplus_server/domain/problemset/service/ProblemSetSaveService.java
@@ -35,44 +35,4 @@ public class ProblemSetSaveService {
 
         return problemSetRepository.save(problemSet).getId();
     }
-
-    @Transactional
-    public void reorderProblems(Long problemSetId, ProblemReorderRequest request) {
-        ProblemSet problemSet = problemSetRepository.findByIdElseThrow(problemSetId);
-
-        // 기존 문항 ID 리스트 업데이트 (순서 반영)
-        List<ProblemId> updatedProblemIds = request.newProblems().stream()
-                .map(ProblemId::new)
-                .collect(Collectors.toList());
-
-        problemSet.updateProblemOrder(updatedProblemIds);
-    }
-
-    @Transactional
-    public void updateProblemSet(Long problemSetId, ProblemSetUpdateRequest request) {
-        ProblemSet problemSet = problemSetRepository.findByIdElseThrow(problemSetId);
-
-        // 문항 리스트 검증
-        List<ProblemId> problemIdList = request.problems().stream()
-                .map(ProblemId::new)
-                .collect(Collectors.toList());
-        problemIdList.forEach(problemRepository::findByIdElseThrow);
-
-        problemSet.updateProblemSet(request.problemSetName(), problemIdList);
-    }
-
-    @Transactional
-    public void deleteProblemSet(Long problemSetId) {
-        ProblemSet problemSet = problemSetRepository.findByIdElseThrow(problemSetId);
-        problemSet.deleteProblemSet();
-    }
-
-    @Transactional
-    public boolean toggleConfirmProblemSet(Long problemSetId) {
-        ProblemSet problemSet = problemSetRepository.findByIdElseThrow(problemSetId);
-
-        // 현재 상태 반전 (true → false, false → true)
-        problemSet.toggleConfirm(!problemSet.isConfirmed());
-        return problemSet.isConfirmed();
-    }
 }
